backend/requirements.txt
torch
numpy
pillow
requests
torchvision
ultralytics
gradio
